\name{manhattanVAAST-package}
\alias{manhattanVAAST-package}
\alias{manhattanVAASTplot}
\docType{package}
\title{
manhattanVAAST
}
\description{
manhattanVAAST plot VAAST pvalues over range.  Similar to GWAS style manhattan plots.
}
\details{
\tabular{ll}{
Package: \tab manhattanVAAST\cr
Type: \tab Package\cr
Version: \tab 1.001\cr
Date: \tab 2012-02-08\cr
License: \tab none \cr
}
}
\author{
Zev N. Kronenberg

Maintainer: <zev.kronenberg@gmail.com>
Lab page: <http://www.yandell-lab.org/software/vaast.html>

}

\references{
A probabilistic disease-gene finder for personal genomes
Yandell M Huff CD Hu H Singleton M Moore B Xing J Jorde L Reese MG
Genome Res. 2011 Jul

Using VAAST to Identify an X-Linked Disorder Resulting in Lethality in Male Infants Due to N-Terminal Acetyltransferase Deficiency
Rope AF Wang K Evjenth R Xing J Johnston JJ Swensen JJ Johnson WJ Moore B Huff CD Bird LM Carey JC Opitz JM Stevens CA Jiang T Schank C Fain HD Robison R Dalley B Chin S South ST Pysher TJ Jorde LB Hakonarson H Lillehaug JR Biesecker LG Yandell M Arnesen T Lyon GJ
Am J Hum Genet. 2011 Jul 15;89(1):28-43
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~

}

\examples{

}
